fix: maintainability hardening — file splits, function extraction, magic numbers
Major fixes (13):
- Split analysis.py stream helpers into analysis_stream.py (299 lines)
- Extract subagent processing into _subagent_runner.py (runner.py → 299 lines)
- Split _weighted_overall into helpers (scoring.py → 294 lines)
- Extract numerical scoring into _scoring_numerical.py (scoring_internals.py → 249 lines)
- Extract MCP args into _mcp_args.py (mcp_findings.py → 283 lines)
- Extract _enrich_judgment/_read_judgments from evidence_parser.py
- Extract _should_respawn/_collect_done from subagent_pool.py run()
- Replace global _fetch_failures with _FetchClient class in knowledge_refresh.py
- Make _build_req_refs_lookup/_resolve_llm_refs public in evidence_parser.py
- Replace duplicate file extraction with shared analysis_stream module
- Extract shared LRU cache and dedup patterns

Minor fixes (49):
- Add module docstrings to ports/dimensions.py, ports/evaluations.py
- Delete empty test_sa_manager.py
- Remove unused import os from test_action_api.py
- Fix relative path in test_readme.py
- Replace magic numbers with named constants across http_client.py,
  dashboard/runner.py, repo_handler.py, report.py
- Add disciplines_conf property to ConfigPaths
- Reorder constants in scaffold.py
- Move threading import to top in knowledge_refresh.py
- Extract zip helpers into action_api_zip.py
- Extract retry loop in standards_fetcher.py, add _parse_asvs_l1 docstring
- Fix hardcoded path in dashboard/cli.py
- Reduce _get_previous_run_for_dimension params with RunLookupCache
- Extract duplicate test helpers to module level
- Split test_mcp_findings.py and test_scoring.py into focused test files
- Lazy env var reads in analysis.py
