+++ src/quodeq/engine/_scoring_numerical.py
@@ -1,0 +1,78 @@
+"""Numerical-mode scoring helpers: deduction computation and grade drops."""
+from __future__ import annotations
+
+# Progressive drop tables: (min_type_count_inclusive, levels_to_drop).
+# Checked top-to-bottom; first matching row wins.
+_CRITICAL_DROP_TABLE: list[tuple[int, int]] = [(12, 3), (4, 2), (1, 1)]
+_MAJOR_DROP_TABLE: list[tuple[int, int]] = [(36, 3), (12, 2), (4, 1)]
+
+# Per-type deduction constants for numerical mode.
+_CRITICAL_PENALTY = 2.0   # per distinct critical violation type, cap 3 types
+_MAJOR_PENALTY = 1.0      # per distinct major violation type, cap 5 types
+_MINOR_PENALTY = 0.25     # per distinct minor violation type
+
+
+def build_deductions(violation_type_counts: dict[str, int], scale_multiplier: int = 1) -> dict:
+    """Compute point deductions for numerical mode.
+
+    Rules:
+    - Each distinct critical violation type removes 2.0 points; types are capped
+      at 3*scale before computing the deduction.
+    - Each distinct major violation type removes 1.0 point; capped at 5*scale.
+    - Minor violation types are NOT capped — every distinct type deducts 0.25.
+    - If the raw critical count reaches 3*scale, the score is hard-capped at 3.
+    - If the raw major count reaches 5*scale, the score is hard-capped at 5.
+    - Both caps may apply simultaneously (take min).
+    """
+    n_critical = violation_type_counts.get("critical", 0)
+    n_major = violation_type_counts.get("major", 0)
+    n_minor = violation_type_counts.get("minor", 0)
+
+    critical_type_cap = 3 * scale_multiplier
+    major_type_cap = 5 * scale_multiplier
+
+    effective_critical = min(n_critical, critical_type_cap)
+    effective_major = min(n_major, major_type_cap)
+
+    critical_deduction = effective_critical * _CRITICAL_PENALTY
+    major_deduction = effective_major * _MAJOR_PENALTY
+    minor_deduction = n_minor * _MINOR_PENALTY
+
+    cap_from_critical = 3 if n_critical >= critical_type_cap else 10
+    cap_from_major = 5 if n_major >= major_type_cap else 10
+
+    return {
+        "critical_type_count": n_critical,
+        "major_type_count": n_major,
+        "minor_type_count": n_minor,
+        "critical_deduction": critical_deduction,
+        "major_deduction": major_deduction,
+        "minor_deduction": minor_deduction,
+        "total_deduction": critical_deduction + major_deduction + minor_deduction,
+        "critical_cap": cap_from_critical,
+        "major_cap": cap_from_major,
+    }
+
+
+def count_grade_drops(violation_type_counts: dict[str, int], scale_multiplier: int = 1) -> int:
+    """Return the number of grade levels to drop in non-numerical mode.
+
+    Drop table thresholds are multiplied by scale_multiplier so that large
+    projects require proportionally more violation types before incurring drops.
+    """
+    n_critical = violation_type_counts.get("critical", 0)
+    n_major = violation_type_counts.get("major", 0)
+
+    critical_drops = 0
+    for min_count, levels in _CRITICAL_DROP_TABLE:
+        if n_critical >= min_count * scale_multiplier:
+            critical_drops = levels
+            break
+
+    major_drops = 0
+    for min_count, levels in _MAJOR_DROP_TABLE:
+        if n_major >= min_count * scale_multiplier:
+            major_drops = levels
+            break
+
+    return max(critical_drops, major_drops)
