+++ src/quodeq/engine/analysis.py
@@ -1,9 +1,4 @@
-"""AI CLI subprocess runner and stream-json evidence extractor.
-
-Spawns the AI CLI with codebase exploration tools (Bash, Glob, Grep, Read),
-captures stream-json output, and extracts JSONL evidence lines.
-Uses an MCP tool server so findings stream in real time via tool calls.
-"""
+"""AI CLI subprocess runner — spawns the AI CLI, captures stream-json, extracts JSONL."""
 from __future__ import annotations
 
 import json
@@ -16,6 +11,12 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Callable
 
+from quodeq.engine.analysis_stream import (
+    count_files_in_stream,
+    count_jsonl_lines,
+    extract_files_from_event,
+    parse_stream_event,
+)
 from quodeq.shared.logging import log_debug, log_warning
 from quodeq.shared.utils import get_ai_cmd, get_ai_model
 
@@ -79,89 +80,26 @@ def _create_mcp_config(
     return Path(tmp.name)
 
 
-def _count_jsonl_lines(jsonl_file: Path) -> int:
-    """Count evidence lines in the JSONL file written by the MCP server."""
-    try:
-        if not jsonl_file.exists():
-            return 0
-        with open(jsonl_file) as f:
-            return sum(1 for line in f if line.strip())
-    except OSError:
-        return 0
-
-
-def _extract_files_from_blocks(blocks: list) -> set[str]:
-    """Extract file paths from Read/Grep tool_use blocks."""
-    files: set[str] = set()
-    for block in blocks:
-        if isinstance(block, dict) and block.get("type") == "tool_use" and block.get("name") in ("Read", "Grep"):
-            fp = (block.get("input") or {}).get("file_path") or (block.get("input") or {}).get("path")
-            if fp:
-                files.add(fp)
-    return files
-
-
-def _parse_stream_event(line: str) -> dict | None:
-    """Parse a single stream event line, returning None for empty or invalid lines."""
-    stripped = line.strip()
-    if not stripped:
-        return None
-    try:
-        return json.loads(stripped)
-    except json.JSONDecodeError:
-        return None
-
-
-def _extract_files_from_event(data: dict) -> set[str]:
-    """Dispatch to the appropriate file extractor based on event type."""
-    etype = data.get("type", "")
-    if etype == "assistant":
-        return _extract_files_from_blocks(data.get("message", {}).get("content", []))
-    if etype == "item.completed":
-        return _extract_files_from_blocks(data.get("item", {}).get("content", []))
-    return set()
-
-
-def _count_files_from_stream(stream_file: Path) -> set[str]:
-    """Extract unique file paths from Read/Grep tool_use events in the stream."""
-    files: set[str] = set()
-    try:
-        with open(stream_file) as f:
-            for line in f:
-                data = _parse_stream_event(line)
-                if data is not None:
-                    files.update(_extract_files_from_event(data))
-    except (OSError, ValueError) as exc:
-        log_debug(f"Failed to count files from stream {stream_file}: {exc}")
-    return files
-
-
-def _count_stream_progress(stream_file: Path, jsonl_file: Path | None = None) -> dict:
-    """Count files read (from stream) and evidence found (from JSONL or stream)."""
-    files = _count_files_from_stream(stream_file)
-    evidence_count = _count_jsonl_lines(jsonl_file) if jsonl_file is not None else 0
-    return {"files_read": len(files), "evidence": evidence_count}
-
-
 def count_files_from_stream(stream_file: Path) -> int:
     """Public: count unique files read by the AI from the stream file."""
-    return len(_count_files_from_stream(stream_file))
+    return len(count_files_in_stream(stream_file))
 
 
-_AI_TOOLS: str = os.environ.get("QUODEQ_AI_TOOLS", "Glob,Grep,Read")
-_BASE_AI_ARGS: tuple[str, ...] = tuple(
-    os.environ.get("QUODEQ_AI_BASE_ARGS", "--print --output-format stream-json --verbose").split()
-)
+_DEFAULT_AI_TOOLS = "Glob,Grep,Read"
+_DEFAULT_BASE_AI_ARGS = "--print --output-format stream-json --verbose"
 
-# Provider-keyed configuration: extend this dict to add support for a new AI
-# provider without modifying _build_ai_cmd or _build_analysis_env.
+
+def _get_ai_tools() -> str:
+    return os.environ.get("QUODEQ_AI_TOOLS", _DEFAULT_AI_TOOLS)
+
+
+def _get_base_ai_args() -> tuple[str, ...]:
+    return tuple(os.environ.get("QUODEQ_AI_BASE_ARGS", _DEFAULT_BASE_AI_ARGS).split())
+
 _PROVIDER_CONFIGS: dict[str, dict] = {
     "claude": {
-        # Extra CLI flags added when an MCP config is present.
         "mcp_permission_args": ["--permission-mode", "bypassPermissions"],
-        # Env vars set when not already present in the subprocess environment.
         "env_set_if_missing": {"CODEX_SANDBOX": "read-only"},
-        # Env vars removed from the subprocess environment.
         "env_remove": ["CLAUDECODE"],
     },
     "codex": {
@@ -177,17 +115,13 @@ class AnalysisError(RuntimeError):
 
 
 def _build_ai_cmd(
-    prompt: str,
-    config: AnalysisConfig,
+    prompt: str, config: AnalysisConfig,
 ) -> tuple[list[str], Path | None]:
-    """Build the AI CLI command line and optional MCP config path.
-
-    Returns (args_list, mcp_config_path_or_None).
-    """
+    """Build the AI CLI command line and optional MCP config path."""
     cmd = config.ai_cmd or get_ai_cmd()
     model = config.ai_model or get_ai_model()
 
-    args = [cmd, *_BASE_AI_ARGS, "--tools", _AI_TOOLS]
+    args = [cmd, *_get_base_ai_args(), "--tools", _get_ai_tools()]
 
     provider_cfg = _PROVIDER_CONFIGS.get(cmd, {})
     mcp_config_path: Path | None = None
@@ -225,6 +159,52 @@ def _terminate_process(process: subprocess.Popen) -> None:
         process.kill()
 
 
+class _IncrementalProgressReader:
+    """Reads new bytes from stream/JSONL files since last check."""
+
+    def __init__(self, stream_file: Path, jsonl_file: Path | None) -> None:
+        self._stream_file = stream_file
+        self._jsonl_file = jsonl_file
+        self._stream_offset = 0
+        self._jsonl_offset = 0
+        self._seen_files: set[str] = set()
+        self._jsonl_count = 0
+
+    def read_progress(self) -> dict:
+        """Return incremental progress since the last call."""
+        self._read_stream()
+        self._read_jsonl()
+        return {"files_read": len(self._seen_files), "evidence": self._jsonl_count}
+
+    def _read_stream(self) -> None:
+        try:
+            with open(self._stream_file, "rb") as f:
+                f.seek(self._stream_offset)
+                new_bytes = f.read()
+                self._stream_offset += len(new_bytes)
+            for line in new_bytes.decode("utf-8", errors="replace").splitlines():
+                data = parse_stream_event(line)
+                if data is not None:
+                    self._seen_files.update(extract_files_from_event(data))
+        except (OSError, ValueError) as exc:
+            log_debug(f"Failed to read stream {self._stream_file}: {exc}")
+
+    def _read_jsonl(self) -> None:
+        if self._jsonl_file is None or not self._jsonl_file.exists():
+            return
+        try:
+            with open(self._jsonl_file, "rb") as jf:
+                jf.seek(self._jsonl_offset)
+                new_bytes = jf.read()
+                self._jsonl_offset += len(new_bytes)
+            self._jsonl_count += sum(
+                1 for line in new_bytes.decode("utf-8", errors="replace").splitlines()
+                if line.strip()
+            )
+        except OSError as exc:
+            log_debug(f"Failed to read JSONL {self._jsonl_file}: {exc}")
+
+
 def _run_with_heartbeat(
     process: subprocess.Popen,
     config: AnalysisConfig,
@@ -234,44 +214,10 @@ def _run_with_heartbeat(
 
     Terminates the process if *max_duration* seconds elapse.
     Returns True if the process was terminated due to timeout.
-
-    Uses an incremental byte-offset reader so each heartbeat only processes
-    new bytes appended since the last read, avoiding repeated full-file scans.
     """
     elapsed = 0
-    max_dur = config.max_duration
     timed_out = False
-    _stream_offset = 0
-    _seen_files: set[str] = set()
-    _jsonl_offset = 0
-    _jsonl_count = 0
-
-    def _read_stream_incremental() -> dict:
-        nonlocal _stream_offset, _jsonl_offset, _jsonl_count
-        try:
-            with open(stream_file, "rb") as f:
-                f.seek(_stream_offset)
-                new_bytes = f.read()
-                _stream_offset += len(new_bytes)
-            for line in new_bytes.decode("utf-8", errors="replace").splitlines():
-                data = _parse_stream_event(line)
-                if data is not None:
-                    _seen_files.update(_extract_files_from_event(data))
-        except (OSError, ValueError) as exc:
-            log_debug(f"Failed to read stream {stream_file}: {exc}")
-        if config.jsonl_file is not None and config.jsonl_file.exists():
-            try:
-                with open(config.jsonl_file, "rb") as jf:
-                    jf.seek(_jsonl_offset)
-                    new_bytes_j = jf.read()
-                    _jsonl_offset += len(new_bytes_j)
-                _jsonl_count += sum(
-                    1 for line in new_bytes_j.decode("utf-8", errors="replace").splitlines()
-                    if line.strip()
-                )
-            except OSError as exc:
-                log_debug(f"Failed to read JSONL {config.jsonl_file}: {exc}")
-        return {"files_read": len(_seen_files), "evidence": _jsonl_count}
+    reader = _IncrementalProgressReader(stream_file, config.jsonl_file)
 
     while process.poll() is None:
         try:
@@ -279,10 +225,9 @@ def _run_with_heartbeat(
         except subprocess.TimeoutExpired:
             elapsed += config.heartbeat_interval
             if config.heartbeat_callback:
-                progress = _read_stream_incremental()
-                config.heartbeat_callback(elapsed, progress)
-            if max_dur is not None and elapsed >= max_dur:
-                log_warning(f"Analysis exceeded max duration ({max_dur}s) — terminating")
+                config.heartbeat_callback(elapsed, reader.read_progress())
+            if config.max_duration is not None and elapsed >= config.max_duration:
+                log_warning(f"Analysis exceeded max duration ({config.max_duration}s) — terminating")
                 _terminate_process(process)
                 timed_out = True
     return timed_out
@@ -327,19 +272,10 @@ def _check_process_result(process: subprocess.Popen, stream_err: Path) -> None:
 
 
 def run_analysis(
-    work_dir: Path,
-    prompt: str,
-    stream_file: Path,
+    work_dir: Path, prompt: str, stream_file: Path,
     config: AnalysisConfig | None = None,
 ) -> None:
-    """Spawn AI CLI subprocess with tools, capturing stream-json to *stream_file*.
-
-    When *config.jsonl_file* is provided, an MCP findings server is configured so
-    the AI reports findings as tool calls that stream directly to the JSONL file.
-
-    Raises:
-        AnalysisError: If the subprocess exits with a non-zero code.
-    """
+    """Spawn AI CLI subprocess, capturing stream-json to *stream_file*."""
     cfg = config or AnalysisConfig()
     args, mcp_config_path = _build_ai_cmd(prompt, cfg)
     env = _build_analysis_env(cfg.ai_cmd or get_ai_cmd())
