--- tests/engine/test_sa_manager.py
@@ -1,2 +1,0 @@
-# Tests for SA manager — removed alongside dead-code stubs (CWE-561).
-# Re-add when SA integration is implemented (issue SA-INTEGRATION).
