+++ tests/test_action_api.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import os
-
 import pytest
 
 from quodeq.action_api import create_app
