+++ tests/test_readme.py
@@ -2,5 +2,6 @@ from pathlib import Path
 
 
 def test_readme_mentions_quodeq_dashboard():
-    text = Path("README.md").read_text()
+    readme = Path(__file__).resolve().parent.parent / "README.md"
+    text = readme.read_text()
     assert "quodeq dashboard" in text
